initial s3 branch commit
Experimental branch that removes API Gateway Integration and uses
two S3 buckets for processing images. The "input" bucket holds
incoming images and uses the "CreateObject" trigger to notify the
memeify lambda that an image is ready for processing. The processed
image is placed in a separate "output" bucket. NOTE: there is currently
no way to specify the text strings to be placed on the top/bottom of
the image. Strings are hardcoded for now...